Created a store backing our Todo List.
Since I got a specification of actions (I created them before), I can
now implement a store responding to those actions. It stores task list
in an immutable list. Since setState from Alt by default mutates, there
is a helper for it, called ImmutableUtil which is built into an Alt
itself. It is a perfect fit for this immutable store use case.

There is a problem with working on store with React-Hot-Loader - since
store cannot be re-initialized in Alt it throws a warning each time I
edit store's implementation. There is a fix for it, very similar in use
to ImmutableUtil - but I have better idea how to achieve it, without
polluting code at all.