Created TodoList and TodoListTask components.
This is a major part of this project. It shows how you can integrate Alt
stores with React components. This is the most verbose way to do it.
There is an utility called `connectToStores` which you can use. I wanted
to show how it is achieved in an explicit way (and to be honest, I
dislike the API of this utility), so I decided to go with an explicit
approach.