Modified main application file to render TodoList component.
For now it is impossible to add new tasks, but it renders an empty state
correctly and has a proper state. You can check it by yourself by
console.logging state in a render method.

To check whether it updates if you call an add action, I'd recommend to
import TodoListActions in this file and call the action in code. I
tested it that way.