Prepared a simple actions for adding new task form.
The logic behind add new task form will be store-based - it seems to be
a rather good pattern to work with Flux-based forms. I want to be able
to change content of a new task and clear the form after I add it.
That's why I prepared two actions for it - 'changeContent' and
'clearForm'.